pyproject.toml
[project]
name = "farminggame"
version = "0.1.0"
description = "Add your description here"
authors = []
dependencies = [
    "pygame-ce>=2.5.2",
]
readme = "README.md"
requires-python = ">= 3.12"

[tool.rye.scripts]
"start" = { cmd="python ./src/__main__.py", desc="Run the game" }

[tool.rye]
managed = false
